+++ egui_demo_lib/src/apps/demo/widget_gallery.rs
@@ -53,9 +53,8 @@ impl super::Demo for WidgetGallery {
 
 impl super::View for WidgetGallery {
     fn ui(&mut self, ui: &mut egui::Ui) {
-        ui.scope(|ui| {
+        ui.add_enabled_ui(self.enabled, |ui| {
             ui.set_visible(self.visible);
-            ui.set_enabled(self.enabled);
 
             egui::Grid::new("my_grid")
                 .num_columns(2)